feat: update delete book by id usecase to use exists book by id gateway

+++ internal/application/usecase/book.go
@@ -1,6 +1,9 @@
 package usecase
 
 import (
+	"errors"
+
+	"github.com/cleysonph/bookshelf-go/internal/application"
 	"github.com/cleysonph/bookshelf-go/internal/application/gateway"
 	"github.com/cleysonph/bookshelf-go/internal/domain"
 )
@@ -38,5 +41,8 @@ func NewDeleteBookUseCase(bookGateway gateway.BookGateway) *DeleteBookUseCase {
 }
 
 func (u *DeleteBookUseCase) Execute(id string) error {
+	if !u.bookGateway.ExistsById(id) {
+		return application.NewBookNotFoundError(errors.New("book not found"), "Book not found")
+	}
 	return u.bookGateway.DeleteById(id)
 }
